// frontend/lib/api/generated/model/solicitudCocinaDetalleOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface SolicitudCocinaDetalleOut {
  solicitud_cocina_detalle_id: number;
  producto_id: number;
  producto_codigo: string;
  producto_nombre: string;
  cantidad_solicitada: number;
  cantidad_teorica_bom: number | null;
}